# Optimize the cost of calling on the main/draw threads