<!-- If this is your first pull request for Fyne please read the contributor docs at:
https://github.com/fyne-io/fyne/wiki/Contributing.
Be sure that your work is based off `develop` branch. --> 

### Description:

Calling a function on the main thread is a common operation
in Go's GUI programming. For instance, the runOnMain must
be called at each render loop.

Previously, the runOnMain allocates a done channel to wait
for a scheduled call on the main thread to complete and thus
can produce the amount of work and awake the garbage collector
quite often. Instead of allocating a channel for each call
on the main thread, this change uses a channel pool and
optimizes this particular case.

Aside, the calls on the draw thread can also be optimized and
use the same channel pool.

Benchmark results (benchstat):

```
$ go test -run=^$ -bench=. -count=10
```

```
goos: linux
goarch: amd64
pkg: fyne.io/fyne/internal/driver/glfw
cpu: Intel(R) Core(TM) i9-9900K CPU @ 3.60GH

name          old time/op     new time/op      delta
RunOnMain-16   5.26µs ±1%      5.02µs ±2%     -4.64%  (p=0.000 n=10+10)
RunOnDraw-16  44.5µs ±12%     43.2µs ±10%        ~    (p=0.393 n=10+10)

name          old alloc/op   new alloc/op      delta
RunOnMain-16    96.0B ±0%       0.0B        -100.00%  (p=0.000 n=10+10)
RunOnDraw-16     107B ±3%        10B ±33%    -90.15%  (p=0.000 n=10+10)

name          old allocs/op new allocs/op      delta
RunOnMain-16     1.00 ±0%      0.00         -100.00%  (p=0.000 n=10+10)
RunOnDraw-16     2.00 ±0%      1.00 ±0%      -50.00%  (p=0.000 n=10+10)
```

Trace visualization:

```
$ cd internal/driver/glfw
& go test -run=none -bench=Main -trace trace.out
```


Before

![image](https://user-images.githubusercontent.com/5498964/105199824-7fcada80-5b3f-11eb-999f-01c42b922181.png)

After:

![image](https://user-images.githubusercontent.com/5498964/105199841-86595200-5b3f-11eb-9fca-48a6edeccea4.png)

(explanation: no garbage!)


### Checklist:
<!-- Please tick these as appropriate using [x] -->

- [x] Tests included.
- [x] Lint and formatter run with no errors.
- [x] Tests all pass.