@andydotxyz 2021-01-26
I was about to approve this (well, after I have tested some of our larger apps) but wanted to check:
Will the pool grow indefinitely if we get huge numbers of simultaneous requests?