@changkun 2021-01-26
> Will the pool grow indefinitely if we get huge numbers of simultaneous requests?

Pool allocated objects will be garbage collected every two GC cycles (after 1.13). In particular, after the first GC cycle, allocated objects are victims and will be reused if there are any needs to reuse them, otherwise, they will be released.

The [GC pacer](https://docs.google.com/document/d/1wmjrocXIWTr1JxU-3EQBI6BK6KgtiFArkG47XK73xIQ/edit#) has a [soft and hard limit](https://github.com/golang/proposal/blob/master/design/14951-soft-heap-limit.md) that provides the progress guarantee, which prevents the growth of the heap goes to infinite.