@andydotxyz 2021-02-06
I'm sorry I forgot. We had a discussion on the community call but I didn't get back to you.