@andydotxyz 2021-02-07
Ah yes, sorry for the ambiguity. This was given the go-ahead. I was waiting until we had forked the bugfix release as this wants to be in main develop for significant testing before release :)